Issue #248 refactor: SearchSchedule 리사이클러 뷰 데이터 타입 변경
페이징을 위해 기존의 방식에서 멀티 뷰 타입으로 변경함.
