+++ app/src/main/java/com/drunkenboys/calendarun/ui/searchschedule/model/SearchScheduleItem.kt
@@ -3,16 +3,36 @@ package com.drunkenboys.calendarun.ui.searchschedule.model
 import androidx.recyclerview.widget.DiffUtil
 import com.drunkenboys.calendarun.data.schedule.entity.Schedule
 import com.drunkenboys.calendarun.util.amPm
+import java.time.LocalDate
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
 
-data class DateScheduleItem(val schedule: Schedule, val onClick: () -> Unit) {
+sealed class SearchScheduleItem
 
-    val duration: String = run {
+data class DateItem(val date: LocalDate) : SearchScheduleItem() {
+
+    override fun toString(): String = date.format(DateTimeFormatter.ofPattern("M월 d일"))
+
+    companion object {
+
+        val diffUtil by lazy {
+            object : DiffUtil.ItemCallback<DateItem>() {
+
+                override fun areItemsTheSame(oldItem: DateItem, newItem: DateItem) = oldItem.hashCode() == newItem.hashCode()
+
+                override fun areContentsTheSame(oldItem: DateItem, newItem: DateItem) = oldItem == newItem
+            }
+        }
+    }
+}
+
+data class DateScheduleItem(val schedule: Schedule, val onClick: () -> Unit) : SearchScheduleItem() {
+
+    override fun toString(): String {
         val startDateFormat = DateTimeFormatter.ofPattern("${schedule.startDate.amPm} hh:mm")
         val endDateFormat = getEndDateFormat(schedule.startDate, schedule.endDate)
 
-        "${schedule.startDate.format(startDateFormat)} ~ ${schedule.endDate.format(endDateFormat)}"
+        return "${schedule.startDate.format(startDateFormat)} ~ ${schedule.endDate.format(endDateFormat)}"
     }
 
     private fun getEndDateFormat(startDate: LocalDateTime, endDate: LocalDateTime): DateTimeFormatter {
